refactor(identifiers): moved `ArrayKey` mod to own subdir, improved organization
