+++ crates/identifiers/src/array_keys/macros.rs
@@ -1,108 +1,7 @@
-//! Utilities for working with identifiers as database keys.
+//! Macros for declaring [`ArrayKey`] impls, along with the impls we declare
+//! with them for the primitive types.
 
-/// Used for types that can be represented as a flat array of bytes, as would be
-/// used for binary database keys.
-///
-/// Unfortunately, `generic_const_exprs` is not stable, so we can't make this
-/// interface *really* nice, we just have to hope that LLVM realizes that it can
-/// eliminate all these bounds checks.
-pub trait ArrayKey {
-    /// The size of an array that contains this key byte.
-    const BYTES: usize;
-
-    /// Copies the key bytes into the array.
-    ///
-    /// This must be big-endian.
-    ///
-    /// # Panics
-    ///
-    /// If called with a slice that is not [`Self::BYTES`] bytes.
-    fn copy_into(&self, buf: &mut [u8]);
-
-    /// Constructs an instance of the key by decoding from bytes.
-    ///
-    /// # Panics
-    ///
-    /// If called with a slice that is not [`Self::BYTES`] bytes.
-    fn copy_from(buf: &[u8]) -> Self;
-}
-
-/// General impl for all bytebufs since they're already of the right format.
-impl<const N: usize> ArrayKey for [u8; N] {
-    const BYTES: usize = N;
-
-    fn copy_into(&self, buf: &mut [u8]) {
-        assert_buf_len_eq::<Self>(buf);
-        buf.copy_from_slice(self);
-    }
-
-    fn copy_from(buf: &[u8]) -> Self {
-        assert_buf_len_eq::<Self>(buf);
-        buf.try_into().unwrap()
-    }
-}
-
-/// Extension trait for "buffer types" that we might want to decode as an [`ArrayKey`].
-pub trait ArrayKeyBuf {
-    /// Attempts to decode the buf into an [`ArrayKey`] type.
-    ///
-    /// Returns `None` if the array is not the right size.
-    fn try_into_key<K: ArrayKey>(&self) -> Option<K>;
-}
-
-impl<T: AsRef<[u8]>> ArrayKeyBuf for T {
-    fn try_into_key<K: ArrayKey>(&self) -> Option<K> {
-        let buf = <Self as AsRef<[u8]>>::as_ref(self);
-        if buf.len() == K::BYTES {
-            Some(K::copy_from(buf))
-        } else {
-            None
-        }
-    }
-}
-
-/// Extension trait for array types that we might want to decode as an [`ArrayKey`].
-///
-/// Unlike [`ArrayKeyBuf`], this is infallible, since the array length is known
-/// statically and MUST be checked at compile time.
-///
-/// You shouldn't have to implement this on your own types, it's just so we can
-/// add trait member fn to `[u8; N]`.
-pub trait ArrayKeyArr<const N: usize> {
-    /// Decodes the array into an [`ArrayKey`] type of exactly the same width.
-    ///
-    /// Instantiating this with a key type whose width isn't `N` MUST fail to
-    /// compile.
-    fn into_key<K: ArrayKey>(self) -> K;
-}
-
-/// Impl for `[u8; N]` arrays that provides the guarantees we expect.
-impl<const N: usize> ArrayKeyArr<N> for [u8; N] {
-    fn into_key<K: ArrayKey>(self) -> K {
-        // Ideally this would be a bound like `K: ArrayKey<BYTES = { N }>`, but
-        // that needs `associated_const_equality`, which isn't stable.  An
-        // inline const inherits the generics of the fn it's in, so we can check
-        // the same thing here, just at monomorphization time instead of when
-        // typechecking the call.
-        assert_key_width_eq::<N, K>();
-        K::copy_from(&self)
-    }
-}
-
-/// Fails to compile if `K` is not exactly `N` bytes wide.
-///
-/// This is a post-monomorphization error, so it only fires where this is
-/// actually instantiated.  That's every real call, but it does mean a bad
-/// instantiation sitting in generic code that's never called goes unreported.
-#[inline(always)]
-fn assert_key_width_eq<const N: usize, K: ArrayKey>() {
-    const {
-        assert!(
-            N == K::BYTES,
-            "arraykey: array length does not match key width"
-        )
-    };
-}
+use super::traits::{ArrayKey, assert_buf_len_eq};
 
 /// Declares an [`ArrayKey`] impl for a tuple of the arity implied by the number
 /// of type parameter names passed to it.
@@ -214,7 +113,7 @@ macro_rules! decl_array_key_struct_impl {
             const BYTES: usize = 0 $(+ <$field_ty as $crate::ArrayKey>::BYTES)+;
 
             fn copy_into(&self, buf: &mut [u8]) {
-                $crate::key_types::assert_buf_len_eq::<Self>(buf);
+                $crate::array_keys::assert_buf_len_eq::<Self>(buf);
                 let mut rest = buf;
                 $(
                     let (cur, next) = core::mem::take(&mut rest)
@@ -226,7 +125,7 @@ macro_rules! decl_array_key_struct_impl {
             }
 
             fn copy_from(buf: &[u8]) -> Self {
-                $crate::key_types::assert_buf_len_eq::<Self>(buf);
+                $crate::array_keys::assert_buf_len_eq::<Self>(buf);
                 let mut rest = buf;
                 $(
                     let (cur, next) = rest.split_at(<$field_ty as $crate::ArrayKey>::BYTES);
@@ -243,22 +142,12 @@ macro_rules! decl_array_key_struct_impl {
 pub(crate) use decl_array_key_struct_impl;
 pub(crate) use decl_array_key_wrapper_impl;
 
-#[inline(always)]
-pub(crate) fn assert_buf_len_eq<K: ArrayKey>(arr: &[u8]) {
-    let len = arr.len();
-    assert_eq!(
-        arr.len(),
-        K::BYTES,
-        "arraykey: passed invalid array (exp {}, got {len})",
-        K::BYTES
-    );
-}
-
 #[cfg(all(test, feature = "arbitrary"))]
 mod tests {
     use arbitrary::{Arbitrary, Unstructured};
 
     use super::*;
+    use crate::array_keys::traits::ArrayKeyArr;
 
     /// Number of instances we check per type.
     const ROUNDS: usize = 64;
